+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This page lists the main changes made to Myokit in each release.
   - [#661](https://github.com/MichaelClerx/myokit/pull/661) When importing MathML, the inverse hyperbolic trig functions are now rendered using slightly simpler equations.
   - [#664](https://github.com/MichaelClerx/myokit/pull/664) EasyML export now adds meta data, a group of CVODE-solved variables, a group of variables to trace, and a group of parameters (based on code contributed by Ed Vigmond).
   - [#664](https://github.com/MichaelClerx/myokit/pull/664) EasyML export now converts voltage, current, and time variables to the preferred units.
+  - [#664](https://github.com/MichaelClerx/myokit/pull/664) EasyML export now has consistently ordered output when re-run.
   - [#664](https://github.com/MichaelClerx/myokit/pull/664) EasyML expression writer now uses `expm1` where possible.
 - Deprecated
   - [#622](https://github.com/MichaelClerx/myokit/pull/622) `SimulationOpenCL.is2d()` was deprecated in favour of `SimulationOpenCL.is_2d()`.